FullStackApp/backend/requirements-minimal.txt
# Core Flask functionality - FAST BUILD
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Session==0.5.0
Flask-Caching==2.0.2
requests==2.31.0
psycopg2-binary==2.9.6
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.2.3

# Essential utilities
pandas==2.0.3
reportlab==4.0.4
xlsxwriter==3.1.2

# Security and monitoring (lightweight)
flask-limiter==3.3.1
argon2-cffi==21.3.0
